Move some folders with manager code related to the apps
