+++ static/skywire-manager-src/src/app/app-routing.module.ts
@@ -11,7 +11,7 @@ import { AppsComponent } from './components/pages/node/apps/apps.component';
 import { SidenavComponent } from './components/layout/sidenav/sidenav.component';
 import { AllTransportsComponent } from './components/pages/node/routing/all-transports/all-transports.component';
 import { AllRoutesComponent } from './components/pages/node/routing/all-routes/all-routes.component';
-import { AllAppsComponent } from './components/pages/node/apps/node-apps/all-apps/all-apps.component';
+import { AllAppsComponent } from './components/pages/node/apps/all-apps/all-apps.component';
 import { NodeInfoComponent } from './components/pages/node/node-info/node-info.component';
 
 const routes: Routes = [
