+++ static/skywire-manager-src/src/app/app.module.ts
@@ -11,28 +11,23 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { FormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCheckboxModule } from '@angular/material/checkbox';
-import { MatChipsModule } from '@angular/material/chips';
 import { ErrorStateMatcher, ShowOnDirtyErrorStateMatcher, RippleGlobalOptions, MAT_RIPPLE_GLOBAL_OPTIONS } from '@angular/material/core';
 import { MAT_DIALOG_DEFAULT_OPTIONS, MatDialogModule } from '@angular/material/dialog';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
-import { MatListModule } from '@angular/material/list';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatProgressBarModule } from '@angular/material/progress-bar';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatSelectModule } from '@angular/material/select';
-import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { MatSnackBarModule, MAT_SNACK_BAR_DEFAULT_OPTIONS } from '@angular/material/snack-bar';
-import { MatTableModule } from '@angular/material/table';
 import { MatTabsModule } from '@angular/material/tabs';
-import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { TransportListComponent } from './components/pages/node/routing/transport-list/transport-list.component';
-import { NodeAppsListComponent } from './components/pages/node/apps/node-apps/node-apps-list/node-apps-list.component';
+import { NodeAppsListComponent } from './components/pages/node/apps/node-apps-list/node-apps-list.component';
 import { CopyToClipboardTextComponent } from './components/layout/copy-to-clipboard-text/copy-to-clipboard-text.component';
 import { ActionsComponent } from './components/pages/node/actions/actions.component';
-import { LogComponent } from './components/pages/node/apps/node-apps/log/log.component';
+import { LogComponent } from './components/pages/node/apps/node-apps-list/log/log.component';
 import { SettingsComponent } from './components/pages/settings/settings.component';
 import { PasswordComponent } from './components/pages/settings/password/password.component';
 import { ClipboardService } from './services/clipboard.service';
@@ -58,13 +53,13 @@ import { ViewAllLinkComponent } from './components/layout/view-all-link/view-all
 import { AllTransportsComponent } from './components/pages/node/routing/all-transports/all-transports.component';
 import { PaginatorComponent } from './components/layout/paginator/paginator.component';
 import { AllRoutesComponent } from './components/pages/node/routing/all-routes/all-routes.component';
-import { AllAppsComponent } from './components/pages/node/apps/node-apps/all-apps/all-apps.component';
+import { AllAppsComponent } from './components/pages/node/apps/all-apps/all-apps.component';
 import { TabBarComponent } from './components/layout/tab-bar/tab-bar.component';
 import { RouteReuseStrategy } from '@angular/router';
 import { AppReuseStrategy } from './app.reuse-strategy';
 import { ConfirmationComponent } from './components/layout/confirmation/confirmation.component';
 import { TransportDetailsComponent } from './components/pages/node/routing/transport-list/transport-details/transport-details.component';
-import { LogFilterComponent } from './components/pages/node/apps/node-apps/log/log-filter/log-filter.component';
+import { LogFilterComponent } from './components/pages/node/apps/node-apps-list/log/log-filter/log-filter.component';
 import { SnackbarComponent } from './components/layout/snack-bar/snack-bar.component';
 import { InitialSetupComponent } from './components/pages/login/initial-setup/initial-setup.component';
 import { SelectLanguageComponent } from './components/layout/select-language/select-language.component';
