+++ static/skywire-manager-src/src/app/components/pages/node/apps/all-apps/all-apps.component.html
