+++ static/skywire-manager-src/src/app/components/pages/node/apps/node-apps-list/log/log.component.ts
