+++ static/skywire-manager-src/src/app/components/pages/node/apps/node-apps-list/node-apps-list.component.html
