+++ static/skywire-manager-src/src/app/components/pages/node/apps/node-apps-list/node-apps-list.component.ts
@@ -3,19 +3,19 @@ import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { Observable, Subscription } from 'rxjs';
 import { ActivatedRoute, Router } from '@angular/router';
 
-import { Application } from '../../../../../../app.datatypes';
-import { AppsService } from '../../../../../../services/apps.service';
-import { LogComponent } from '../log/log.component';
-import { NodeComponent } from '../../../node.component';
-import { AppConfig } from '../../../../../../app.config';
-import GeneralUtils from '../../../../../../utils/generalUtils';
-import { ConfirmationComponent } from '../../../../../layout/confirmation/confirmation.component';
-import { SnackbarService } from '../../../../../../services/snackbar.service';
+import { Application } from '../../../../../app.datatypes';
+import { AppsService } from '../../../../../services/apps.service';
+import { LogComponent } from './log/log.component';
+import { NodeComponent } from '../../node.component';
+import { AppConfig } from '../../../../../app.config';
+import GeneralUtils from '../../../../../utils/generalUtils';
+import { ConfirmationComponent } from '../../../../layout/confirmation/confirmation.component';
+import { SnackbarService } from '../../../../../services/snackbar.service';
 import { SelectableOption, SelectOptionComponent } from 'src/app/components/layout/select-option/select-option.component';
-import { SkysocksSettingsComponent } from '../skysocks-settings/skysocks-settings.component';
+import { SkysocksSettingsComponent } from '../node-apps/skysocks-settings/skysocks-settings.component';
 import { processServiceError } from 'src/app/utils/errors';
 import { OperationError } from 'src/app/utils/operation-error';
-import { SkysocksClientSettingsComponent } from '../skysocks-client-settings/skysocks-client-settings.component';
+import { SkysocksClientSettingsComponent } from '../node-apps/skysocks-client-settings/skysocks-client-settings.component';
 import { TranslateService } from '@ngx-translate/core';
 import { FilterKeysAssociation, FilterTextElements, filterList, updateFilterTexts } from 'src/app/utils/filters';
 import {
